Updated models so the database-setup is functional Added Controller for all new models correspond to the database design - Tested functionality through the python-script. Everything seems to work.
Two things to note:
The timestamp is converted from string to date through a predefined pattern - Might be a smarter way to do it. But it works

Second, in the BatchReportController routes to find humidities, temps etc for a specific batch is implemented. Replacing the findAll route in the corresponding controllers.